% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_families.R
\docType{data}
\name{gene_families}
\alias{gene_families}
\alias{gene_families_full}
\title{Gene families}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 3159 rows and 4 columns.

An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 3209 rows and 6 columns.
}
\usage{
gene_families

gene_families_full
}
\description{
Mostly transcribed from O. Hobert, Wormbook, 2013: http://www.wormbook.org/chapters/www_neuronalgenome/neuronalgenome.html
}
\details{
May need refinement. There are two files, gene_families_full, a more direct transcription of Hobert (2013), and
gene_families, a simplified version.
}
\keyword{datasets}
