% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genes_by_pattern.R
\docType{data}
\name{genes_by_pattern}
\alias{genes_by_pattern}
\title{Genes with particular expression patterns}
\format{
## `gene_by_pattern`
A list of length 7:
\describe{
  \item{nondetected}{Note detected in scRNA-Seq of L4 hermaphrodites.}
  \item{nonneuronal}{detected in cells that are not neurons.}
  \item{ubiquitous}{likely expressed in every cell type.}
  \item{panneuronal}{Likely expressed in every neuron, and only in neurons.}
  \item{broadly_neuronal}{Expressed in many neurons.}
  \item{neuron_specific}{Expressed in neurons, and no other cell type.}
  \item{present_in_neurons}{The combination of the above lists.}
}
}
\usage{
genes_by_pattern
}
\description{
Based on scRNA-Seq (10x genomics generated by CeNGEN), lists genes with particular expression patterns, e.g. panneuronal etc
}
\keyword{datasets}
