% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbp_matia-gonzalez_2015.R
\docType{data}
\name{list_rbp_matiagonzalez2015}
\alias{list_rbp_matiagonzalez2015}
\title{List of RNA-binding proteins}
\format{
## `list_rbp_matiagonzalez2015`
A dataframe with 3 columns:
\describe{
  \item{selected}{1 if it's ine of the 594 genes selected (see paper)}
  \item{gene_id}{Wormbase Gene ID}
  \item{gene_name}{Gene symbol as of WS289}
}
}
\usage{
list_rbp_matiagonzalez2015
}
\description{
Table S3b of [Ana M Matia-González, Emma E Laing & André P Gerber](https://www.nature.com/articles/nsmb.3128).
}
\details{
Conserved mRNA-binding proteomes in eukaryotic organisms
Nature Structural & Molecular Biology volume 22, pages 1027–1033 (2015)
}
\keyword{datasets}
