% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbp_tamburino2013.R
\docType{data}
\name{list_rbp_tamburino2013}
\alias{list_rbp_tamburino2013}
\title{List of RNA-binding proteins}
\format{
## `list_rbp_tamburino2013`
A list of length 8:
\describe{
  \item{gene_name}{Gene symbol.}
  \item{gene_id}{Wormbase ID as of WS219.}
  \item{ORF}{Sequence name.}
  \item{RBD}{RNA binding domain.}
  \item{Group}{Gene specificity, see paper.}
  \item{Source}{Source.}
  \item{GO}{GO or not.}
  \item{UniProtKB}{UniProtKB or not}
}
}
\usage{
list_rbp_tamburino2013
}
\description{
Table S2 of [Tamburino, Ryder and Walhout, G3, 2013](https://doi.org/10.1534/g3.112.004390).
}
\details{
A Compendium of Caenorhabditis elegans RNA Binding Proteins Predicts Extensive Regulation at Multiple Levels
G3 Genes|Genomes|Genetics, Volume 3, Issue 2, 1 February 2013, Pages 297–304, https://doi.org/10.1534/g3.112.004390
}
\keyword{datasets}
