% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuron_properties.R
\docType{data}
\name{neuron_properties}
\alias{neuron_properties}
\title{Various properties of neuron classes}
\format{
## `neuron_properties`
A data frame 128 × 27:
\describe{
  \item{Neuron_type}{Neuron name (e.g. ADL, AIY, ...)}
  \item{Modality}{Interneuron, motor neuron, sensory neuron, or pharyngeal neuron.}
  \item{Neurotransmitter}{Neurotransmitter (can be several separated by comma)}
  \item{Ciliated, Pharyngeal, Nb_Processes, Process_Position, Branching, Sexual_dimorphism, Mechanosensation, Gentle_Touch, Harsh_Touch, Nociception, Proprioception, Chemosensation, Thermosensation, Photosensation, Electrosensation, Texture, Oxygen_sensory, CO2_sensory, Odorsensation, Pheromone_sensory, Magnetic_Field, Humidity, Osmosensation, Dauer_Entry}{Various properties.}
}
}
\usage{
neuron_properties
}
\description{
Data collected by Alec Barrett and others.
}
\keyword{datasets}
